Modification de personnageList
