+++ src/components/PageBio.jsx
@@ -20,7 +20,7 @@ export class pageBio extends React.Component {
                         <ButtonSettings />              
                     </Grid>
                 </Grid>
-                <PersonnageList character={this.props.character} />
+                <PersonnageList setPersoOne={this.props.setPersoOne} persoOne={this.props.persoOne} setPersoTwo={this.props.setPersoTwo} persoTwo={this.props.persoTwo} apiCharacters={this.props.apiCharacters} />
             </div>
         )
     }
